requirements.txt
streamlit==1.44.0
llama-index==0.12.25
llama-index-core==0.12.25
llama-index-embeddings-huggingface==0.5.2
llama-index-llms-gemini==0.4.13
pandas==2.2.3
openpyxl==3.1.5
huggingface-hub==0.29.3
sentence-transformers==4.0.0
faiss-cpu==1.10.0
tiktoken==0.9.0
docx2txt==0.9
PyMuPDF==1.25.4
lxml==5.3.1
python-dotenv==1.1.0
scikit-learn==1.6.1
numpy==2.2.4
google-generativeai==0.8.4
google-auth==2.38.0
requests==2.32.3
